mcp/scripts/requirements.txt
openai>=1.0.0
psycopg[binary]>=3.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
pgvector>=0.2.0
